fix compile error

+++ cocos/scripting/lua-bindings/manual/lua_cocos2dx_manual.cpp
@@ -5085,8 +5085,8 @@ static int lua_cocos2dx_Label_createWithTTF01(lua_State* L)
             {
                 luaval_to_size(L, 5, &dimensions);
             }
-            TextHAlignment hAlignment = static_cast<TextHAlignment>(tolua_tonumber(L, 6, 0));
-            TextVAlignment vAlignment = static_cast<TextVAlignment>(tolua_tonumber(L, 7, 0));
+            TextHAlignment hAlignment = static_cast<TextHAlignment>((int)tolua_tonumber(L, 6, 0));
+            TextVAlignment vAlignment = static_cast<TextVAlignment>((int)tolua_tonumber(L, 7, 0));
             
             cocos2d::Label* ret = cocos2d::Label::createWithTTF(text, fontFile, fontSize, dimensions, hAlignment, vAlignment);
             
@@ -5137,8 +5137,8 @@ static int lua_cocos2dx_Label_create_deprecated(lua_State* tolua_S)
             {
                 luaval_to_size(tolua_S, 5, &dimensions);
             }
-            TextHAlignment hAlignment = static_cast<TextHAlignment>(tolua_tonumber(tolua_S, 6, 0));
-            TextVAlignment vAlignment = static_cast<TextVAlignment>(tolua_tonumber(tolua_S, 7, 0));
+            TextHAlignment hAlignment = static_cast<TextHAlignment>((int)tolua_tonumber(tolua_S, 6, 0));
+            TextVAlignment vAlignment = static_cast<TextVAlignment>((int)tolua_tonumber(tolua_S, 7, 0));
             
             cocos2d::Label* ret = cocos2d::Label::create(text, fontFile, fontSize, dimensions, hAlignment, vAlignment);
